@@ -4,11 +4,10 @@ import { RecipesComponent } from './recipes/recipes.component';
 import { ShoppingListComponent } from './shopping-list/shopping-list.component';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
-  standalone: true,
-  imports: [HeaderComponent, RecipesComponent, ShoppingListComponent],
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.css'],
+    imports: [HeaderComponent, RecipesComponent, ShoppingListComponent]
 })
 export class AppComponent {
   loadedFeature = signal('recipe');
