@@ -1,27 +1,20 @@
-import { Component, inject, OnInit, signal } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { RecipeListComponent } from './recipe-list/recipe-list.component';
 import { RecipeDetailComponent } from './recipe-detail/recipe-detail.component';
-import { Recipe } from './recipe.model';
+
 import { RecipeService } from './recipe.service';
 
 @Component({
-  selector: 'app-recipes',
-  templateUrl: './recipes.component.html',
-  styleUrls: ['./recipes.component.css'],
-  standalone: true,
-  imports: [RecipeListComponent, RecipeDetailComponent],
+    selector: 'app-recipes',
+    templateUrl: './recipes.component.html',
+    styleUrls: ['./recipes.component.css'],
+    imports: [RecipeListComponent, RecipeDetailComponent]
 })
 export class RecipesComponent implements OnInit {
-  selectedRecipe = signal<Recipe>(undefined);
-  private rService = inject(RecipeService);
+  private recipeService = inject(RecipeService);
+  selectedRecipe = this.recipeService.recipeSelected;
 
   constructor() {}
 
-  ngOnInit() {
-    this.rService.recipeSelected.subscribe({
-      next: (recipe: Recipe) => {
-        this.selectedRecipe.set(recipe);
-      }
-    });
-  }
+  ngOnInit() {}
 }
